upgrade gradle, fixed up renderer, changed min java to 11 and added basic jetbrain annotations

+++ src/main/java/com/mrcrayfish/vehicle/entity/BoatEntity.java
@@ -110,7 +110,7 @@ public abstract class BoatEntity extends PoweredVehicleEntity
                     if(fluidState.is(FluidTags.WATER))
                     {
                         float waterLevel = (float) y + fluidState.getHeight(this.level, pooledMutable);
-                        this.waterLevel = Math.max((double) waterLevel, this.waterLevel);
+                        this.waterLevel = Math.max(waterLevel, this.waterLevel);
                         inWater |= boundingBox.minY < (double) waterLevel;
                     }
                 }
